Split AndroidProject into BaseAndroidProject, Installable and Startable
