+++ src/main/scala/AndroidTestProject.scala
@@ -1,11 +1,7 @@
 import sbt._
 
-abstract class AndroidTestProject(info: ProjectInfo) extends AndroidProject(info) {
+abstract class AndroidTestProject(info: ProjectInfo) extends BaseAndroidProject(info) with Installable {
   override def proguardInJars = Path.emptyPathFinder
-
-  // Normally it makes no sense to start a test application, so make these actions no-ops.
-  override def startEmulatorAction = task { None }
-  override def startDeviceAction = task { None }
 
   lazy val testEmulator = instrumentationTestAction(true) describedAs("runs tests in emulator") dependsOn reinstallEmulator
   lazy val testDevice = instrumentationTestAction(false) describedAs("runs tests on device") dependsOn reinstallDevice
